+++ tests/integration/cases/call_scalar_args/Scala_heterogeneous_strategy_record_call@v3.scala
@@ -1,0 +1,6 @@
+object Fixture_call_scalar_args_Scala_heterogeneous_strategy_record_call {
+def process(value: Any = null): Any = null
+process(value = "hello")
+process(value = 42)
+process(value = true)
+}