+++ tests/integration/cases/record_basic/Scala_heterogeneous_strategy_record@v3.scala
@@ -1,0 +1,13 @@
+object Fixture_record_basic_Scala_heterogeneous_strategy_record {
+case class Record0(id: Int, description: String, is_done: Boolean, blocks: List[Int])
+val my_data = Record0(
+    id = 1,
+    description = "She said \"hello\", then waved",
+    is_done = false,
+    blocks = List[Int](
+        1,
+        2,
+        3,
+    ),
+)
+}